Add ca64 assembler, improved errors + .BYTE support
This adds a dependency on thiserror for defining a custom
AssembleError type, for tokenizing, parsing or IO errors.

For the sake of passing the required lifetime needed for
VerboseErrors then parsing code is now also given a reference
to the original input.

The .BYTE directive can also be used via a .DB alias which
some other assemblers support.

This ended up removing support for labels without a
colon but I'm not sure what the usecase for those was.